separate VC and View
